Bump patch version to 0.1.4
due to successive tries to publish the package onto Atmosphere.

Turns out there was a simple timeout problem, fixed by increasing that timeout with command `TIMEOUT_SCALE_FACTOR=10 meteor publish`. See https://github.com/meteor/meteor/pull/7143 and https://forums.meteor.com/t/1-3-4-breaks-galaxy-deployment-etimedout/25383/17
But in the first place, it is VERY strange that Meteor needs to build and upload a few MB for such a simple package… mys:fonts package does not need to download such a big payload when installed on a project. To be investigated later on. Might be a change in Meteor behaviour.